Basic Slidebar Implemented soft sliding with ponderation
I have got some trouble getting the mosue position in the HTML
relative to the div.

It has jumps, right now there is a ponderation to achive the smoth movement

There is an interpolation function due to is impossible to get 0 to 180 angle
only using the mause possition.

Like the calculus are: (X-1.84)-14 = Y
1.84 = (180/157)
157 = 157-14